+++ SomnusBackend/src/main/java/com/somnus/server/backend/config/auth/firebase/FirebaseAuthenticationProvider.java
@@ -1,0 +1,51 @@
+package com.somnus.server.backend.config.auth.firebase;
+
+import com.google.api.core.ApiFuture;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseToken;
+import com.somnus.server.backend.users.domain.User;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.authentication.dao.AbstractUserDetailsAuthenticationProvider;
+import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.userdetails.UserDetails;
+import org.springframework.security.web.authentication.session.SessionAuthenticationException;
+import org.springframework.stereotype.Component;
+
+import java.util.concurrent.ExecutionException;
+
+/**
+ * @author Anthony Waters
+ * https://github.com/awaters1/spring-security-firebase/tree/46ff34ce0d44bb3728420046b6d1f20fa8ef4f8b
+ */
+
+@Component
+public class FirebaseAuthenticationProvider extends AbstractUserDetailsAuthenticationProvider {
+
+	@Autowired
+	private FirebaseAuth firebaseAuth;
+
+	@Override
+	public boolean supports(Class<?> authentication) {
+		return (FirebaseAuthenticationToken.class.isAssignableFrom(authentication));
+	}
+
+	@Override
+	protected void additionalAuthenticationChecks(UserDetails userDetails,
+			UsernamePasswordAuthenticationToken authentication) throws AuthenticationException {
+	}
+
+	@Override
+	protected UserDetails retrieveUser(String username, UsernamePasswordAuthenticationToken authentication)
+			throws AuthenticationException {
+		final FirebaseAuthenticationToken authenticationToken = (FirebaseAuthenticationToken) authentication;
+
+		ApiFuture<FirebaseToken> task = firebaseAuth.verifyIdTokenAsync(authenticationToken.getToken());
+		try {
+			FirebaseToken token = task.get();
+			return new User(token.getEmail(), token.getUid());
+		} catch (InterruptedException | ExecutionException e) {
+			throw new SessionAuthenticationException(e.getMessage());
+		}
+	}
+}