+++ SomnusBackend/src/main/java/com/somnus/server/backend/emailApi/MailSender.java
@@ -1,4 +1,0 @@
-package com.somnus.server.backend.emailApi;
-
-public class MailSender {
-}