+++ SomnusBackend/src/main/java/com/somnus/server/backend/teammembers/TeamMemberService.java
@@ -2,8 +2,11 @@ package com.somnus.server.backend.teammembers;
 
 import com.somnus.server.backend.exceptions.ErrorMessage;
 import com.somnus.server.backend.exceptions.SomnusException;
+import com.somnus.server.backend.teammembers.database.ContributionRepository;
 import com.somnus.server.backend.teammembers.database.TeamMemberRepository;
+import com.somnus.server.backend.teammembers.domain.Contribution;
 import com.somnus.server.backend.teammembers.domain.TeamMember;
+import com.somnus.server.backend.teammembers.dto.ContributionDto;
 import com.somnus.server.backend.teammembers.dto.TeamMemberDto;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.retry.annotation.Backoff;
@@ -23,6 +26,9 @@ public class TeamMemberService {
     @Autowired
     private TeamMemberRepository teamMemberRepository;
 
+    @Autowired
+    private ContributionRepository contributionRepository;
+
     @Retryable(
             value = {SQLException.class},
             backoff = @Backoff(delay = 5000))
@@ -30,7 +36,8 @@ public class TeamMemberService {
     public List<TeamMemberDto> getAllTeamMembers() {
         List<TeamMemberDto> teamMemberDtos = new ArrayList<>();
         for(TeamMember teamMember : teamMemberRepository.findAll()) {
-            teamMemberDtos.add(new TeamMemberDto());
+            teamMemberDtos.add(new TeamMemberDto(teamMember.getTeamMemberName(), teamMember.getPhotoPath(),
+                    teamMember.getDateJoined(), teamMember.getNumContributions()));
         }
         return teamMemberDtos;
     }
@@ -41,13 +48,13 @@ public class TeamMemberService {
     @Transactional(isolation = Isolation.READ_COMMITTED)
     public TeamMemberDto getTeamMember(Integer id) {
         Optional<TeamMember> optionalTeamMember = teamMemberRepository.findById(id);
-        if(!optionalTeamMember.isPresent()) {
+        if(optionalTeamMember == null || !optionalTeamMember.isPresent()) {
             throw new SomnusException(ErrorMessage.NO_TEAMMEMBER_FOUND);
         }
 
         TeamMember teamMember = optionalTeamMember.get();
         return new TeamMemberDto(teamMember.getTeamMemberName(), teamMember.getPhotoPath(),
-                teamMember.getDateJoined());
+                teamMember.getDateJoined(), teamMember.getNumContributions());
     }
 
     @Retryable(
@@ -65,5 +72,23 @@ public class TeamMemberService {
     @Transactional(isolation = Isolation.READ_COMMITTED)
     public void deleteTeamMember(Integer teamMemberId) {
         teamMemberRepository.deleteById(teamMemberId);
+    }
+
+    @Retryable(
+            value = {SQLException.class},
+            backoff = @Backoff(delay = 5000))
+    @Transactional(isolation = Isolation.READ_COMMITTED)
+    public void addContribution(ContributionDto contributionDto) {
+        Optional<TeamMember> optionalTeamMember = teamMemberRepository.findById(contributionDto.getTeamMemberId());
+        if(!optionalTeamMember.isPresent()) {
+            throw new SomnusException(ErrorMessage.NO_TEAMMEMBER_FOUND);
+        }
+
+        TeamMember teamMember = optionalTeamMember.get();
+        contributionRepository.save(new Contribution(contributionDto.getTitle(), teamMember,
+                contributionDto.getDescription()));
+
+        teamMember.addContribution();
+        teamMemberRepository.save(teamMember);
     }
 }