+++ SomnusBackend/src/main/java/com/somnus/server/backend/users/domain/Role.java
@@ -1,0 +1,5 @@
+package com.somnus.server.backend.users.domain;
+
+public enum Role {
+    ADMIN, EDITOR, READER
+}